refactor!: remove deprecated GeneratedVaadinRichTextEditor
